Update maximum/minimum atom canonicalization
