+++ src/atoms/affine.rs
@@ -226,6 +226,32 @@ pub(crate) fn broadcast_to(expr: Expr, expr_shape: &Shape, target_shape: &Shape)
     broadcast_2d_to(expr, expr_shape, target_shape)
 }
 
+pub(crate) fn broadcast_elementwise_exprs(
+    exprs: impl IntoIterator<Item = Expr>,
+) -> (Shape, Vec<Expr>) {
+    let exprs: Vec<Expr> = exprs.into_iter().collect();
+    let target_shape = exprs
+        .iter()
+        .map(|expr| expr.shape())
+        .reduce(|acc, shape| {
+            acc.broadcast(&shape)
+                .unwrap_or_else(|| panic!("cannot broadcast shapes {} and {}", acc, shape))
+        })
+        .expect("elementwise atom requires at least one expression");
+
+    let exprs = exprs
+        .into_iter()
+        .map(|expr| {
+            let expr_shape = expr.shape();
+            broadcast_to(expr, &expr_shape, &target_shape).unwrap_or_else(|| {
+                panic!("cannot broadcast shape {} to {}", expr_shape, target_shape)
+            })
+        })
+        .collect();
+
+    (target_shape, exprs)
+}
+
 fn broadcast_2d_to(expr: Expr, expr_shape: &Shape, target_shape: &Shape) -> Option<Expr> {
     if !expr_shape.is_matrix() || !target_shape.is_matrix() {
         return None;
